requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
arrow==1.2.3
    # via django-q
asgiref==3.5.2
    # via django
backports-zoneinfo==0.2.1
    # via django
blessed==1.19.1
    # via django-q
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
coreapi==2.3.3
    # via
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via coreapi
dj-database-url==1.0.0
    # via -r requirements.in
dj-stripe==2.7.2
    # via -r requirements.in
django==4.1.2
    # via
    #   -r requirements.in
    #   dj-database-url
    #   dj-stripe
    #   django-cors-headers
    #   django-picklefield
    #   django-q
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   jsonfield
django-cors-headers==3.13.0
    # via -r requirements.in
django-picklefield==3.1
    # via django-q
django-q==1.3.9
    # via -r requirements.in
django-rest-swagger==2.2.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   django-rest-swagger
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jsonfield==3.1.0
    # via dj-stripe
markupsafe==2.1.1
    # via jinja2
numpy==1.23.4
    # via
    #   -r requirements.in
    #   pandas
openapi-codec==1.3.2
    # via django-rest-swagger
pandas==1.5.1
    # via -r requirements.in
pyjwt==2.6.0
    # via djangorestframework-simplejwt
python-dateutil==2.8.2
    # via
    #   arrow
    #   pandas
pytz==2022.5
    # via
    #   djangorestframework
    #   pandas
pyyaml==6.0
    # via -r requirements.in
redis==3.5.3
    # via django-q
requests==2.28.1
    # via
    #   coreapi
    #   stripe
simplejson==3.17.6
    # via django-rest-swagger
six==1.16.0
    # via
    #   blessed
    #   python-dateutil
sqlparse==0.4.3
    # via django
stripe==4.2.0
    # via dj-stripe
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   coreapi
urllib3==1.26.12
    # via requests
wcwidth==0.2.5
    # via blessed

# The following packages are considered to be unsafe in a requirements file:
# setuptools
